+++ src/main/java/pl/thedeem/intellij/common/Icons.java
@@ -1,11 +1,14 @@
 package pl.thedeem.intellij.common;
 
 import com.intellij.openapi.util.IconLoader;
-import com.intellij.ui.LayeredIcon;
 import com.intellij.util.IconUtil;
+import com.intellij.util.ui.JBUI;
+import com.intellij.util.ui.UIUtil;
 import org.jetbrains.annotations.NotNull;
 
 import javax.swing.*;
+import java.awt.*;
+import java.awt.image.BufferedImage;
 
 public interface Icons {
     Icon DYNATRACE_LOGO = IconLoader.getIcon("/icons/dynatrace.svg", Icons.class);
@@ -21,15 +24,41 @@ public interface Icons {
 
     static Icon scaleToBottomRight(@NotNull Icon base, @NotNull Icon original, float scale) {
         Icon scaled = IconUtil.scale(original, null, scale);
+        int badgeX = base.getIconWidth() - scaled.getIconWidth();
+        int badgeY = base.getIconHeight() - scaled.getIconHeight();
+        int border = JBUI.scale(1);
 
-        return new LayeredIcon(2) {{
-            setIcon(base, 0);
-            setIcon(
-                    scaled,
-                    1,
-                    base.getIconWidth() - scaled.getIconWidth(),
-                    base.getIconHeight() - scaled.getIconHeight()
-            );
-        }};
+        return new Icon() {
+            @Override
+            public void paintIcon(Component c, Graphics g, int x, int y) {
+                BufferedImage image = UIUtil.createImage(c, getIconWidth(), getIconHeight(), BufferedImage.TYPE_INT_ARGB);
+                Graphics2D g2d = image.createGraphics();
+                try {
+                    base.paintIcon(c, g2d, 0, 0);
+                    g2d.setComposite(AlphaComposite.Clear);
+                    g2d.fillOval(
+                            badgeX - border,
+                            badgeY - border,
+                            scaled.getIconWidth() + 2 * border,
+                            scaled.getIconHeight() + 2 * border
+                    );
+                    g2d.setComposite(AlphaComposite.SrcOver);
+                    scaled.paintIcon(c, g2d, badgeX, badgeY);
+                } finally {
+                    g2d.dispose();
+                }
+                g.drawImage(image, x, y, null);
+            }
+
+            @Override
+            public int getIconWidth() {
+                return base.getIconWidth();
+            }
+
+            @Override
+            public int getIconHeight() {
+                return base.getIconHeight();
+            }
+        };
     }
 }